refactor(ws-client): fold style fixes, move errors test to tests/, rename keepalive-shared to close-codes
